#432 stage 1 of 3: find the remapping hardware, read what it can do, say
so.  Changes nothing about how any device reaches memory, and exists so
that one sentence stops being an assumption -- a userspace driver on a
machine with no IOMMU can reach all of physical memory, and until now
the kernel could not tell you which kind of machine it was on.

Two readers, DMAR and IVRS, behind one vendor-neutral description; the
second did not change a field of the first, which is the only thing that
tests the shape.  Each engine's registers are then followed and read,
because a misparsed table produces plausible numbers by construction --
real bytes from a real table, read one field over -- and only the
silicon can say the address was right.

Three things came out of the runs rather than the plan:

  KVM, run against this suite for the first time, failed two MSI
  self-tests on every board.  A processor's store to its own local APIC
  page is a register access and not a message, so #457's doorbell test
  was measuring the emulator.  Both known disagreements between the
  accelerators are TCG being permissive, and a permissive emulator does
  not fail -- it passes.  #516.

  Every generated MIG prototype was claiming sizes nobody needs, and one
  generator defect was not cosmetic: arg_descriptor was counted by one
  rule and emitted by another, so every routine after the first named a
  descriptor belonging to some other argument.  #514, closed.

  Warnings to zero on both targets: i386 97 -> 0, x86-64 13 -> 0, on
  clean full builds.

Green under both accelerators and on i386: 170 default, 172 q35, 176 at
-smp 4, 180 with either IOMMU, 68 PASS on the i386 smoke.

Stages 2 and 3 remain, and carry the three DMA clauses inherited from
#457.

+++ uros/src/lib/libmach/mach_init.c
@@ -144,7 +144,14 @@ static int		mach_init(void)
         *  with the code below.
 	* This hack should disappear when gdb improves.
 	*/
-	if ((int)user_data.user_data == MACH_GDB_RUN_MAGIC_NUMBER) {
+	/*
+	 * ⚠️ uintptr_t and not int.  This compared the LOW HALF of a pointer
+	 * against the magic number on x86-64, so any address whose bottom
+	 * thirty-two bits happened to be 1 would have suspended the task --
+	 * rare, silent, and indistinguishable from a debugger attaching.
+	 * gcc said so; nothing else could have.
+	 */
+	if ((uintptr_t)user_data.user_data == MACH_GDB_RUN_MAGIC_NUMBER) {
 	    kern_return_t ret;
 	    user_data.user_data = 0;
